requirements.txt
requests
git+https://github.com/hvnsweeting/uds@master
numpy
pandas
matplotlib
plotly
kaleido
pyyaml
